refactor(registry): update registry interface to support set of constructors as value type
Modified the IRegistry interface and implementations to support Set<TConstructor<T>> as a value
type in addition to the existing T and TConstructor<T> types. This change enables
registering multiple implementations for the same token.

Removed the injectable registry as it's no longer needed with the new approach. The
Injectable decorator has been simplified to just add metadata without registering the class.

Improved error messages in container class and added more context to registry errors.